+++ ContentPatcher/Framework/PatchLoader.cs
@@ -14,7 +14,9 @@ using ContentPatcher.Framework.Tokens.Json;
 using Newtonsoft.Json.Linq;
 using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
+using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
+using StardewValley;
 
 namespace ContentPatcher.Framework
 {
@@ -470,14 +472,19 @@ namespace ContentPatcher.Framework
                             if (fromAsset == null)
                                 return TrackSkip($"must set the {nameof(PatchConfig.FromFile)} field for a {PatchType.Load} patch.");
 
+                            // parse priority
+                            if (!this.TryParsePriority(entry, AssetLoadPriority.Exclusive, out AssetLoadPriority priority, out string? error))
+                                return TrackSkip(error);
+
                             // save
                             patch = new LoadPatch(
                                 indexPath: indexPath,
                                 path: path,
                                 assetName: targetAsset!,
+                                priority: priority,
+                                updateRate: updateRate,
                                 conditions: conditions,
                                 localAsset: fromAsset,
-                                updateRate: updateRate,
                                 contentPack: pack,
                                 parentPatch: parentPatch,
                                 parseAssetName: this.ParseAssetName
@@ -515,21 +522,23 @@ namespace ContentPatcher.Framework
                             List<EditDataPatchField>? fields = null;
                             List<EditDataPatchMoveRecord>? moveEntries = null;
                             List<IManagedTokenString>? targetField = null;
-                            if (fromAsset == null)
-                            {
-                                if (!TryParseFields(tokenParser.Context, entry, out entries, out fields, out moveEntries, out targetField, out string? error))
-                                    return TrackSkip(error);
-                            }
+                            if (fromAsset == null && !TryParseFields(tokenParser.Context, entry, out entries, out fields, out moveEntries, out targetField, out string? error))
+                                return TrackSkip(error);
+
+                            // parse priority
+                            if (!this.TryParsePriority(entry, AssetEditPriority.Default, out AssetEditPriority priority, out error))
+                                return TrackSkip(error);
 
                             // parse text operations
-                            if (!this.TryParseTextOperations(entry, tokenParser, immutableRequiredModIDs, path.With(nameof(entry.TextOperations)), out IList<ITextOperation> textOperations, out string? parseError))
-                                return TrackSkip(parseError);
+                            if (!this.TryParseTextOperations(entry, tokenParser, immutableRequiredModIDs, path.With(nameof(entry.TextOperations)), out IList<ITextOperation> textOperations, out error))
+                                return TrackSkip(error);
 
                             // save
                             patch = new EditDataPatch(
                                 indexPath: indexPath,
                                 path: path,
                                 assetName: targetAsset!,
+                                priority: priority,
                                 conditions: conditions,
                                 fromFile: fromAsset,
                                 records: entries,
@@ -569,11 +578,16 @@ namespace ContentPatcher.Framework
                             if (entry.ToArea != null && !this.TryParseRectangle(entry.ToArea, tokenParser, immutableRequiredModIDs, path.With(nameof(entry.ToArea)), out error, out toArea))
                                 return TrackSkip(error);
 
+                            // parse priority
+                            if (!this.TryParsePriority(entry, AssetEditPriority.Default, out AssetEditPriority priority, out error))
+                                return TrackSkip(error);
+
                             // save
                             patch = new EditImagePatch(
                                 indexPath: indexPath,
                                 path: path,
                                 assetName: targetAsset!,
+                                priority: priority,
                                 conditions: conditions,
                                 fromAsset: fromAsset,
                                 fromArea: fromArea,
@@ -678,8 +692,8 @@ namespace ContentPatcher.Framework
                             }
 
                             // parse text operations
-                            if (!this.TryParseTextOperations(entry, tokenParser, immutableRequiredModIDs, path.With(nameof(entry.TextOperations)), out IList<ITextOperation> textOperations, out string? parseError))
-                                return TrackSkip(parseError);
+                            if (!this.TryParseTextOperations(entry, tokenParser, immutableRequiredModIDs, path.With(nameof(entry.TextOperations)), out IList<ITextOperation> textOperations, out error))
+                                return TrackSkip(error);
 
                             // read from/to asset areas
                             TokenRectangle? fromArea = null;
@@ -698,11 +712,16 @@ namespace ContentPatcher.Framework
                             if (fromAsset == null && !mapProperties.Any() && !mapTiles.Any() && !addWarps.Any() && !textOperations.Any())
                                 return TrackSkip($"must specify at least one of {nameof(entry.AddWarps)}, {nameof(entry.FromFile)}, {nameof(entry.MapProperties)}, {nameof(entry.MapTiles)}, or {nameof(entry.TextOperations)}");
 
+                            // parse priority
+                            if (!this.TryParsePriority(entry, AssetEditPriority.Default, out AssetEditPriority priority, out error))
+                                return TrackSkip(error);
+
                             // save
                             patch = new EditMapPatch(
                                 indexPath: indexPath,
                                 path: path,
                                 assetName: targetAsset!,
+                                priority: priority,
                                 conditions: conditions,
                                 fromAsset: fromAsset,
                                 fromArea: fromArea,
@@ -757,6 +776,56 @@ namespace ContentPatcher.Framework
             {
                 return TrackSkip($"error reading info. Technical details:\n{ex}");
             }
+        }
+
+        /// <summary>Parse the priority field for a patch.</summary>
+        /// <typeparam name="TPriority">The priority type (one of <see cref="AssetEditPriority"/> or <see cref="AssetLoadPriority"/>).</typeparam>
+        /// <param name="patch">The patch whose priority to parse.</param>
+        /// <param name="defaultValue">The default priority if not specified.</param>
+        /// <param name="priority">The parsed priority value, if valid.</param>
+        /// <param name="error">The error message indicating why parsing failed, if applicable.</param>
+        /// <returns>Returns whether parsing succeeded.</returns>
+        private bool TryParsePriority<TPriority>(PatchConfig patch, TPriority defaultValue, out TPriority priority, [NotNullWhen(false)] out string? error)
+            where TPriority : struct
+        {
+            // default if omitted
+            if (string.IsNullOrWhiteSpace(patch.Priority))
+            {
+                priority = defaultValue;
+                error = null;
+                return true;
+            }
+
+            // parse as enum value
+            if (Utility.TryParseEnum(patch.Priority, out TPriority newPriority))
+            {
+                priority = newPriority;
+                error = null;
+                return true;
+            }
+
+            // parse as an offset like 'Medium + 10'
+            int splitAt = patch.Priority.IndexOfAny(new[] { '-', '+' });
+            if (splitAt > 0)
+            {
+                string rawPriority = patch.Priority[..splitAt];
+                char rawSign = patch.Priority[splitAt];
+                string rawOffset = patch.Priority[(splitAt + 1)..];
+
+                if (Utility.TryParseEnum(rawPriority, out newPriority) && int.TryParse(rawOffset, out int offset))
+                {
+                    if (rawSign == '-')
+                        offset *= -1;
+
+                    priority = (TPriority)(object)((int)(object)newPriority + offset);
+                    error = null;
+                    return true;
+                }
+            }
+
+            priority = defaultValue;
+            error = $"the {nameof(patch.Priority)} value '{patch.Priority}' is invalid: expected one of [{string.Join(", ", Enum.GetNames(typeof(TPriority)))}] or a simple offset like '{Activator.CreateInstance(typeof(TPriority))} + 10'";
+            return false;
         }
 
         /// <summary>Parse the text operation fields for an <see cref="PatchType.EditData"/> or <see cref="PatchType.EditMap"/> patch.</summary>