+++ ContentPatcher/ModEntry.cs
@@ -66,7 +66,8 @@ namespace ContentPatcher
             new Migration_1_27(),
             new Migration_1_28(),
             new Migration_1_29(),
-            new Migration_1_30()
+            new Migration_1_30(),
+            new Migration_2_0()
         };
 
         /// <summary>The special validation logic to apply to assets affected by patches.</summary>